backend/admin_exchange_service/src/test/java/com/cg/stockmarket/admin_exchange_service/exception/GlobalExceptionHandlerTest.java
package com.cg.stockmarket.admin_exchange_service.exception;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.MockitoAnnotations;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;

import static org.junit.jupiter.api.Assertions.assertEquals;


public class GlobalExceptionHandlerTest {

    @InjectMocks
    private GlobalExceptionHandler globalExceptionHandler;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    public void testHandleNotFoundException() {
        NotFoundException exception = new NotFoundException("Exchange not found");
        ResponseEntity<String> response = globalExceptionHandler.handleNotFoundException(exception);

        assertEquals(HttpStatus.NOT_FOUND, response.getStatusCode());
        assertEquals("Exchange not found", response.getBody());
    }

    @Test
    public void testHandleGeneralException() {
        Exception exception = new Exception("Simulated exception");
        ResponseEntity<String> response = globalExceptionHandler.handleGeneralException(exception);

        assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, response.getStatusCode());
        assertEquals("An error occurred: Simulated exception", response.getBody());
    }
}
